feat(bangumi): add type convert mapper
